modify: Firestore 스키마 타입 파일 생성, react-hook-form 데이터 타입 파일 생성 및 import 경로 수정
