+++ src/components/ui/productcard/horizontalcard/HorizontalCard.Style.tsx
@@ -9,7 +9,9 @@ export const CardContainer = styled.div`
   border-radius: ${(props) => props.theme.color.radius};
   align-items: center;
   gap: 25px;
-  box-shadow: rgba(99, 99, 99, 0.2) 0px 2px 8px 0px;
+  box-shadow:
+    rgba(0, 0, 0, 0.05) 0px 6px 24px 0px,
+    rgba(0, 0, 0, 0.08) 0px 0px 0px 1px;
 `;
 
 export const CardImageBox = styled.div<{ $src: string }>`
