+++ src/pages/buyer/home/useHome.ts
@@ -1,6 +1,6 @@
-import { ProductSchema } from '@/firebase';
 import { getValidCategories } from '@/services/categoryService';
 import { fetchProducts } from '@/services/productService';
+import { ProductSchema } from '@/types/FirebaseType';
 import { useQueries, useQuery } from '@tanstack/react-query';
 import { orderBy, where } from 'firebase/firestore';
 import { useEffect, useState } from 'react';
